mainPot styles
